+++ src/lib/pushers/page-pusher/push-pages.ts
@@ -1,0 +1,77 @@
+import * as mgmtApi from "@agility/management-sdk";
+import { state, getApiClient } from "../../../core/state";
+import { PusherResult } from "../../../types/sourceData";
+import { SitemapHierarchy } from "../../shared/sitemap-hierarchy";
+import { PageMapper } from "../../mappers/page-mapper";
+import { processSitemap } from "./process-sitemap";
+
+export async function pushPages(
+	sourceData: mgmtApi.PageItem[],
+	targetData: mgmtApi.PageItem[],
+	// onProgress?: PusherProgressCallback
+): Promise<PusherResult> {
+	// Extract data from sourceData - unified parameter pattern
+	let pages: mgmtApi.PageItem[] = sourceData || [];
+
+	const { sourceGuid, targetGuid, locale } = state;
+	const pageMapper = new PageMapper(sourceGuid[0], targetGuid[0], locale[0]);
+
+	if (!pages || pages.length === 0) {
+		console.log("No pages found to process.");
+		return { status: "success", successful: 0, failed: 0, skipped: 0 };
+	}
+
+	const sitemapHierarchy = new SitemapHierarchy();
+
+	const sitemaps = await sitemapHierarchy.loadAllSitemaps();
+	const channels = Object.keys(sitemaps);
+
+	console.log(`Processing ${pages.length} pages across ${channels.length} channels...`);
+
+	let successful = 0;
+	let failed = 0;
+	let skipped = 0; // No duplicates to skip since API prevents true duplicates at same hierarchy level
+	let status: "success" | "error" = "success";
+	let publishableIds: number[] = []; // Track target page IDs for auto-publishing
+
+
+	//loop all the channels
+	for (const channel of channels) {
+		const sitemap = sitemaps[channel];
+
+		const { sourceGuid, targetGuid, locale, overwrite } = state;
+		const apiClient = getApiClient();
+
+		try {
+			const res = await processSitemap({
+				channel,
+				pageMapper,
+				sitemapNodes: sitemap,
+				sourceGuid: sourceGuid[0],
+				targetGuid: targetGuid[0],
+				locale: locale[0],
+				apiClient,
+				overwrite,
+				sourcePages: pages,
+				// Top-level pages have no parent
+				parentPageID: -1
+			})
+
+			successful = res.successful;
+			failed = res.failed;
+			skipped = res.skipped;
+			publishableIds = res.publishableIds;
+
+			if (failed > 0) {
+				status = "error";
+			}
+
+		} catch (error) {
+			console.error(`⚠️ Error in page processing for channel: ${channel}`, error);
+			status = "error";
+		}
+
+	}
+
+	return { status, successful, failed, skipped, publishableIds };
+}
